暂时告一段落
未能加载到本地

+++ src/main/java/com/linstudy/ttsx/vo/DataVO.java
@@ -9,6 +9,7 @@ public class DataVO<T> {
     // json
     private Integer code;
     private String msg;
-    private Integer count;
+//    private Integer count;
+    private long count;
     private List<T> data;
 }
